fix: bootstrap with containers
