+++ services/user-service/internal/bootstrap/bootstrap.go
@@ -2,7 +2,10 @@ package bootstrap
 
 import (
 	firebaseauth "4it428-newsletter-api/services/user-service/internal/infrastructure/firebase"
+	"4it428-newsletter-api/services/user-service/internal/infrastructure/persistence/repositories"
 	"4it428-newsletter-api/services/user-service/internal/service/auth"
+	"4it428-newsletter-api/services/user-service/internal/service/services"
+	"4it428-newsletter-api/services/user-service/internal/transport/api/v1/handler"
 	"context"
 	firebase "firebase.google.com/go/v4"
 	"github.com/jackc/pgx/v5/pgxpool"
@@ -10,6 +13,8 @@ import (
 	"log"
 	"os"
 )
+
+// SETUP METHODS:
 
 func SetupDatabase(ctx context.Context) (*pgxpool.Pool, error) {
 	// Initialize the database connection pool.
@@ -42,4 +47,37 @@ func SetupFirebaseAuth(ctx context.Context) (auth.IAuthProvider, error) {
 		return nil, err
 	}
 	return authProvider, nil
+}
+
+// CONTAINERS FOR EASIER DI:
+
+type HandlersContainer struct {
+	UserHandler *handler.UserHandler
+	AuthHandler *handler.AuthHandler
+}
+
+func NewHandlersContainer(s *ServicesContainer) *HandlersContainer {
+	return &HandlersContainer{
+		UserHandler: handler.NewUserHandler(s.UserService),
+		AuthHandler: handler.NewAuthHandler(s.AuthService),
+	}
+}
+
+type ServicesContainer struct {
+	UserService services.IUserService
+	AuthService services.IAuthService
+}
+
+func NewServicesContainer(
+	userRepository *repositories.UserRepository,
+	authProvider auth.IAuthProvider,
+) *ServicesContainer {
+
+	userService := services.NewUserService(authProvider, userRepository)
+	authService := services.NewAuthService(authProvider, userService)
+
+	return &ServicesContainer{
+		UserService: userService,
+		AuthService: authService,
+	}
 }
